Clean-up storage Package
