+++ src/webserver/go.mod
@@ -4,6 +4,7 @@ go 1.17
 
 require (
 	github.com/gorilla/mux v1.8.0
+	github.com/joho/godotenv v1.4.0
 	github.com/lib/pq v1.10.4
 	github.com/mattn/go-sqlite3 v1.14.0
 	github.com/stretchr/testify v1.7.0
