+++ src/webserver/storage/database_old.go
@@ -7,6 +7,7 @@ import (
 	"io/ioutil"
 	"log"
 	"net/http"
+	"os"
 	"strconv"
 	"time"
 
@@ -19,8 +20,8 @@ type M map[string]interface{}
 
 var (
 	// Configuration
-	DATABASE = "../minitwitcopy.db"
-	Db      *sql.DB = ConnectDb()
+	DATABASE         = "../minitwitcopy.db"
+	Db       *sql.DB = ConnectDb()
 	UserM    M
 	Session  map[string]string = make(map[string]string)
 )
@@ -31,7 +32,18 @@ const (
 
 // ConnectDb returns a new connection to the database
 func ConnectDb() *sql.DB {
-	return ConnectPsql()
+	connStr := fmt.Sprintf("postgres://%v:%v@%v:%v/%v?sslmode=disable",
+		os.Getenv("POSTGRES_USER"),
+		os.Getenv("POSTGRES_PASSWORD"),
+		"database",
+		5432,
+		os.Getenv("POSTGRES_DB"))
+	db, err := sql.Open("postgres", connStr)
+
+	if err != nil {
+		log.Fatalf("psql.go/ConnectPsql(): Failed to connect to PSQL: %s", err)
+	}
+	return db
 }
 
 // InitDb creates the database tables
